requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
python-multipart==0.0.9
chromadb==0.5.5
fastmcp==2.10.6
mcp>=1.2.0
jinja2==3.1.4
